+++ lib/my/my_strcat.c
@@ -11,7 +11,7 @@ char *my_strcat(char *dest, char const *src)
 {
     int length = my_strlen(dest);
 
-    for (uint64_t i = 0; i < my_strlen(src); i++)
+    for (int i = 0; i < my_strlen(src); i++)
         dest[length + i] = src[i];
     dest[length + my_strlen(src)] = '\0';
     return (dest);
